pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "AI agents repository using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-google-genai>=2.0",
    "python-dotenv>=1.0",
    # Scheduler Agent dependencies
    "slack-bolt>=1.18",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["agents*", "shared*"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
